+++ modules/tinymce/src/core/test/ts/browser/content/InsertContentCommandTest.ts
@@ -66,7 +66,7 @@ describe('browser.tinymce.core.content.InsertContentCommandTest', () => {
     rng.setEnd(editor.getBody(), 0);
     editor.selection.setRng(rng);
     editor.execCommand('mceInsertContent', false, 'x');
-    assert.equal(editor.getContent(), '<p>x</p><hr />');
+    assert.equal(editor.getContent(), '<p>x</p><hr>');
   });
 
   it('mceInsertContent HR at end of H1', () => {
@@ -76,7 +76,7 @@ describe('browser.tinymce.core.content.InsertContentCommandTest', () => {
     editor.execCommand('mceInsertContent', false, '<hr>');
     LegacyUnit.equalDom(editor.selection.getNode(), editor.getBody().lastChild);
     assert.equal(editor.selection.getNode().nodeName, 'H1');
-    assert.equal(editor.getContent(), '<h1>abc</h1><hr /><h1>\u00a0</h1>');
+    assert.equal(editor.getContent(), '<h1>abc</h1><hr><h1>\u00a0</h1>');
   });
 
   it('mceInsertContent HR at end of H1 with P sibling', () => {
@@ -86,7 +86,7 @@ describe('browser.tinymce.core.content.InsertContentCommandTest', () => {
     editor.execCommand('mceInsertContent', false, '<hr>');
     LegacyUnit.equalDom(editor.selection.getNode(), editor.getBody().lastChild);
     assert.equal(editor.selection.getNode().nodeName, 'P');
-    assert.equal(editor.getContent(), '<h1>abc</h1><hr /><p>def</p>');
+    assert.equal(editor.getContent(), '<h1>abc</h1><hr><p>def</p>');
   });
 
   it('mceInsertContent HR at end of H1 with inline elements with P sibling', () => {
@@ -96,7 +96,7 @@ describe('browser.tinymce.core.content.InsertContentCommandTest', () => {
     editor.execCommand('mceInsertContent', false, '<hr>');
     LegacyUnit.equalDom(editor.selection.getNode(), editor.getBody().lastChild);
     assert.equal(editor.selection.getNode().nodeName, 'P');
-    assert.equal(editor.getContent(), '<h1><strong>abc</strong></h1><hr /><p>def</p>');
+    assert.equal(editor.getContent(), '<h1><strong>abc</strong></h1><hr><p>def</p>');
   });
 
   it('mceInsertContent empty block', () => {
@@ -260,7 +260,7 @@ describe('browser.tinymce.core.content.InsertContentCommandTest', () => {
     rng.setEnd(editor.dom.select('p')[0].firstChild, 1);
     editor.selection.setRng(rng);
     editor.execCommand('mceInsertContent', false, '<img src="about:blank" />');
-    assert.equal(editor.getContent(), '<p><img src="about:blank" /></p>');
+    assert.equal(editor.getContent(), '<p><img src="about:blank"></p>');
     rng = normalizeRng(editor.selection.getRng());
     assert.isTrue(rng.collapsed);
     assert.equal(rng.startContainer.nodeName, 'P');
@@ -294,7 +294,7 @@ describe('browser.tinymce.core.content.InsertContentCommandTest', () => {
     rng.setEnd(editor.dom.select('p')[0].firstChild, 2);
     editor.selection.setRng(rng);
     editor.execCommand('mceInsertContent', false, '<hr />');
-    assert.equal(editor.getContent(), '<p>1</p><hr /><p>3</p>');
+    assert.equal(editor.getContent(), '<p>1</p><hr><p>3</p>');
     rng = normalizeRng(editor.selection.getRng());
     assert.isTrue(rng.collapsed);
     LegacyUnit.equalDom(rng.startContainer, editor.getBody().lastChild);